backend/pyproject.toml
[project]
name = "universal-backend"
version = "0.1.0"
description = "Universal Backend for MVP Startups (Django)"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "djangorestframework-simplejwt>=2.3",
    "psycopg[binary]>=3.1",
]
